### 2016-09-19 #74 @karthik
@arfon Since I was one of the reviewers for this particular paper on the rOpenSci side, it would be a COI to be the handling editor on this side for this paper as well. So I am going to assign a new editor.
