### 2016-09-19 #74 @karthik
@acabunoc For context, this paper has already been through a stringent review at rOpenSci (https://github.com/ropensci/onboarding/issues/56). If this submission looks ok to you, just hit accept. No need to send out for review unless you feel something is off with the process that it has already been through.

**More detailed explanation:** We recently introduced a mechanism where someone can 'onboard' a package at rOpenSci and also check a couple more boxes to have this as a paper at JOSS. We check then check JOSS requirements as part of the process so when it comes here, everything is already in place. A human editor just needs to take a quick look and hit accept.

In this _particular_ case, I was also the reviewer there so I cannot hit accept. That's why I've assigned you. So no need for this to go out to review unless you feel that it is not up to JOSS standards.

cc @arfon
