### 2016-09-19 #74 @arfon
> Thanks! I'll take a look ✨

@acabunoc great thanks. Take a look at https://github.com/ropensci/onboarding/issues/56 and see what you think. We should probably make the formal review issue anyway (even for tracking). Also I want to try out @whedon's new super-powers :grin:.

In practice this means:

1. Assigning a reviewer (`@whedon assign @username as reviewer`)
2. Starting the review (`@whedon start review`)
