### 2016-09-19 #74 @acabunoc
Just read through https://github.com/ropensci/onboarding/issues/56 -- looks great and thorough and complies with our own review process. I'll make a formal issue with myself as a reviewer and I'll document the decision there. Let me know if I should go about this differently!

Do I have to use a new comment to to use @whedon's super powers? I'm going to try here:
