### 2016-09-19 #75 @whedon
[REVIEW]: ezknitr: Avoid the typical working directory pain when using 'knitr'
**Submitting author:** @daattali (<a href="http://orcid.org/0000-0002-5645-3493">Dean Attali</a>)
**Repository:** <a href="https://github.com/ropenscilabs/ezknitr" target ="_blank">https://github.com/ropenscilabs/ezknitr</a>
**Version:** v0.5
**Archive:** <a href="http://dx.doi.org/10.6084/m9.figshare.3545363" target="_blank">10.6084/m9.figshare.3545363</a>
