### 2016-09-19 #75 @acabunoc
I hope it's okay that I'm both editor and reviewer!

Based on the review completed for ropensci: https://github.com/ropensci/onboarding/issues/56, I think we should accept this to JOSS
