### 2016-09-19 #75 @arfon
> I hope it's okay that I'm both editor and reviewer!

👍 sounds good to me! Lemme just compile the paper to check things look good there. (@whedon can't do this just yet)
