### 2016-09-19 #75 @arfon
@acabunoc - there was a small update required here: https://github.com/ropenscilabs/ezknitr/pull/9

@daattali - @acabunoc has recommended acceptance here. At this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
