### 2016-09-19 #75 @daattali
rOpenSci already required us to submit to figshare: 10.6084/m9.figshare.3545363
