### 2016-09-20 #75 @arfon
Thanks for handling this @acabunoc.

@daattali - your paper is now accepted into JOSS. Your DOI is http://joss.theoj.org/papers/10.21105/joss.00075 🎉 🚀 💥
